json working

+++ projeto3-parent/rest/shopapi/dbinfo/models.py
@@ -10,7 +10,7 @@ class Instance(models.Model):
     id = models.AutoField(primary_key=True)
     name = models.CharField(max_length=100)
     data_type = models.CharField(max_length=10)
-    revenue = models.DecimalField(max_digits=5, decimal_places=2, default=0)
-    expenses = models.DecimalField(max_digits=5, decimal_places=2, default=0)
-    profit = models.DecimalField(max_digits=5, decimal_places=2, default=0)
+    revenue = models.FloatField(default=0)
+    expenses = models.FloatField(default=0)
+    profit = models.FloatField(default=0)
